+++ src/client/src/containers/Footer/index.tsx
@@ -26,7 +26,7 @@ import { openPostGenModalAction } from "../../store/modals/action";
 
 import { InjectedIntlProps, injectIntl } from "react-intl";
 
-import { getIsVisitedRoutesSelector } from "../../store/templates/wizardContent/wizardNavigationSelector";
+import { getIsVisitedRoutesSelector } from "../../store/templates/preview/wizardNavigationSelector";
 import { isEnableNextPageSelector } from "../../store/selection/app/wizardSelectionSelector/wizardSelectionSelector";
 import { AppState } from "../../store/combineReducers";
