+++ src/client/src/store/config/combineReducers.ts
@@ -8,4 +8,4 @@ const wizardContentReducer = combineReducers({
 });
 
 export default wizardContentReducer;
-export type WizardContentType = ReturnType<typeof wizardContentReducer>;
+export type TemplateType = ReturnType<typeof wizardContentReducer>;
