+++ src/client/src/store/templates/preview/action.ts
