+++ src/client/src/store/templates/preview/model.ts
