+++ src/client/src/store/templates/preview/previewReducer.ts
