+++ src/client/src/store/templates/preview/wizardNavigationSelector.ts
