+++ src/client/src/store/templates/wizardInfoActionType.ts
@@ -5,7 +5,7 @@ import {
 } from "../generationStatus/model";
 
 import { ISetVisitedPageAction, IResetVisitedPageAction, ISetPageAction, ISetDetails } from "./pages/model";
-import { IResetWizardAction } from "./wizardContent/model";
+import { IResetWizardAction } from "./preview/model";
 
 type WizardInfoType =
   | IVersionData
