+++ src/client/src/store/typeKeys.ts
@@ -14,17 +14,15 @@ export enum WIZARD_INFO_TYPEKEYS {
 
 export enum WIZARD_CONTENT_TYPEKEYS {
   GET_VSCODE_API = "WTS/vscode/GET_VSCODE_API",
-  GET_PAGES_OPTIONS = "WTS/wizardContent/GET_PAGES_OPTIONS",
-  GET_PAGES_OPTIONS_SUCCESS = "WTS/wizardContent/GET_PAGES_OPTIONS_SUCCESS",
-  GET_PROJECT_TYPES = "WTS/wizardContent/GET_PROJECT_TYPES",
-  GET_PROJECT_TYPES_SUCCESS = "WTS/wizardContent/GET_PROJECT_TYPES_SUCCESS",
-  GET_WEB_APP_OPTIONS = "WTS/wizardContent/GET_WEB_APP_OPTIONS",
-  GET_WEB_APP_OPTIONS_SUCCESS = "WTS/wizardContent/GET_WEB_APP_OPTIONS_SUCCESS",
-  GET_FRONTEND_FRAMEWORKS = "WTS/wizardContent/GET_FRONTEND_FRAMEWORKS",
-  SET_FRONTEND_FRAMEWORKS = "WTS/wizardContent/SET_FRONTEND_FRAMEWORKS",
-  GET_BACKEND_FRAMEWORKS = "WTS/wizardContent/GET_BACKEND_FRAMEWORKS",
-  SET_PREVIEW_STATUS = "WTS/wizardContent/SET_PREVIEW_STATUS",
-  SET_BACKEND_FRAMEWORKS = "WTS/wizardContent/SET_BACKEND_FRAMEWORKS",
-  UPDATE_FRAMEWORK = "WTS/wizardContent/UPDATE_FRAMEWORK",
-  LOAD_WIZARD_CONTENT = "WTS/wizardContent/LOAD_WIZARD_CONTENT"
+  GET_PAGES_OPTIONS = "WTS/templates/GET_PAGES_OPTIONS",
+  GET_PAGES_OPTIONS_SUCCESS = "WTS/templates/GET_PAGES_OPTIONS_SUCCESS",
+  GET_WEB_APP_OPTIONS = "WTS/templates/GET_WEB_APP_OPTIONS",
+  GET_WEB_APP_OPTIONS_SUCCESS = "WTS/templates/GET_WEB_APP_OPTIONS_SUCCESS",
+  GET_FRONTEND_FRAMEWORKS = "WTS/templates/GET_FRONTEND_FRAMEWORKS",
+  SET_FRONTEND_FRAMEWORKS = "WTS/templates/SET_FRONTEND_FRAMEWORKS",
+  GET_BACKEND_FRAMEWORKS = "WTS/templates/GET_BACKEND_FRAMEWORKS",
+  SET_PREVIEW_STATUS = "WTS/templates/SET_PREVIEW_STATUS",
+  SET_BACKEND_FRAMEWORKS = "WTS/templates/SET_BACKEND_FRAMEWORKS",
+  UPDATE_FRAMEWORK = "WTS/templates/UPDATE_FRAMEWORK",
+  LOAD_WIZARD_CONTENT = "WTS/templates/LOAD_WIZARD_CONTENT"
 }
